# backend/openapi.yaml
openapi: 3.0.3
info:
  title: STS Admin Backend API
  version: 1.0.0
  description: OpenAPI specification for the STS admin backend.
servers:
  - url: /
paths:
...
  /api/admin/auth/login:
    post:
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login succeeded
        "401":
          description: Invalid credentials
...
  /api/complaints/{id}:
    get:
      summary: Get complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complaint returned
...